[冠志]history UI finished
